last update

+++ controller/employee_controller.php
@@ -7,21 +7,27 @@
         $emp_data = json_decode(decode($_POST['new_employee']));
         $fullname = input_data($emp_data->fullname);
         $gender = input_data($emp_data->gender);
-        $dep_id = input_data($emp_data->dep_id);
+        $user_group_id = input_data($emp_data->user_group_id);
         $username = $emp_data->username;
-        $password = password_hash($emp_data->password,true);
-        $address = ($emp_data->address=="")?null:input_data($emp_data->address);
-        $tel = ($emp_data->tel=="")?null:input_data($emp_data->tel);
-        $user_type = input_data($emp_data->user_type);
-        $date_of_birth = ($emp_data->date_of_birth=="")?null:input_data($emp_data->date_of_birth);
-        $sql = "INSERT INTO users (fullname, gender, date_of_birth, address, tel, dep_id, username, password, user_type, status)
-        VALUES(?,?,?,?,?,?,?,?,?,?)";
-        $query = $dbcon->prepare($sql);
-        $query->execute(array($fullname,$gender,$date_of_birth,$address,$tel,$dep_id,$username,$password,$user_type,1));
-        if($query){
-            echo "Swal.fire({icon:'success',html:'<span class=phetsarath>ບັນທຶກຂໍ້ມູນສໍາເລັດ!</span>',allowOutsideClick: false}).then((result) => {if (result.isConfirmed) {window.location.href='template?page=employee'}});";
+        $password = password_hash($emp_data->password,PASSWORD_DEFAULT);
+        $phone_number = ($emp_data->phone_number=="")?null:input_data($emp_data->phone_number);
+        $sql = "SELECT CASE WHEN (SELECT COUNT(*) FROM tb_user WHERE fullname=?)=1 THEN 'ຊື່ພະນັກງານຊໍ້າກັນ' 
+                WHEN (SELECT COUNT(*) FROM tb_user WHERE username=?)=1 THEN 'ຊື່ເຂົ້າໃຊ້ລະບົບຊໍ້າກັນ' 
+                ELSE FALSE END AS result";
+        $return_data = $dbcon->prepare($sql);
+        $return_data->execute(array($fullname,$username));
+        $result = $return_data->fetch(PDO::FETCH_ASSOC)['result'];
+        if(!$result){
+            $sql = "INSERT INTO tb_user(fullname, gender, phone_number, user_group_id, username, password) VALUES (?,?,?,?,?,?)";
+            $query = $dbcon->prepare($sql);
+            $query->execute(array($fullname,$gender,$phone_number,$user_group_id,$username,$password));
+            if($query){
+                echo "Swal.fire({icon:'success',html:'<span class=phetsarath>ບັນທຶກຂໍ້ມູນສໍາເລັດ!</span>',allowOutsideClick: false}).then((result) => {if (result.isConfirmed) {window.location.href='template?page=employee'}});";
+            }else{
+                echo "Swal.fire({icon:'error',html:'<span class=phetsarath>ບັນທຶກຂໍ້ມູນບໍ່ສໍາເລັດ<br>ເກີດຂໍ້ຜິດພາດລະຫວ່າງການບັນທຶກ!</span>'})";
+            }
         }else{
-            echo "Swal.fire({icon:'error',html:'<span class=phetsarath>ບັນທຶກຂໍ້ມູນບໍ່ສໍາເລັດ<br>ເກີດຂໍ້ຜິດພາດລະຫວ່າງການບັນທຶກ!</span>'})";
+            echo "Swal.fire({icon:'error',html:'<span class=phetsarath>".$result."!</span>'})";
         }
     }
     if(isset($_POST['update_employee'])){
@@ -36,13 +42,13 @@
         $dep_id_param = input_data($emp_data->dep_id_param);
         $username = $emp_data->username;
         $password = $emp_data->password;
-        $password_hash = password_hash($emp_data->password,true);
+        $password_hash = password_hash($emp_data->password,PASSWORD_DEFAULT);
         $address = ($emp_data->address=="")?null:input_data($emp_data->address);
         $tel = ($emp_data->tel=="")?null:input_data($emp_data->tel);
         $user_type = input_data($emp_data->user_type);
         $date_of_birth = ($emp_data->date_of_birth=="")?null:input_data($emp_data->date_of_birth);
         if($password==""||$password=="edlquiz"){
-            $sql = "UPDATE users SET fullname=?,gender=?,date_of_birth=?,address=?,tel=?,dep_id=?,username=?,user_type=? WHERE id=?";
+            $sql = "UPDATE tb_user SET fullname=?,gender=?,date_of_birth=?,address=?,tel=?,dep_id=?,username=?,user_type=? WHERE id=?";
             $query = $dbcon->prepare($sql);
             $query->execute(array($fullname,$gender,$date_of_birth,$address,$tel,$dep_id,$username,$user_type,$id));
             if($query){
@@ -51,7 +57,7 @@
                 echo "Swal.fire({icon:'error',html:'<span class=phetsarath>ບັນທຶກຂໍ້ມູນບໍ່ສໍາເລັດ<br>ເກີດຂໍ້ຜິດພາດລະຫວ່າງການບັນທຶກ!</span>'})";
             }
         }else{
-            $sql = "UPDATE users SET fullname=?,gender=?,date_of_birth=?,address=?,tel=?,dep_id=?,username=?,password=?,user_type=? WHERE id=?";
+            $sql = "UPDATE tb_user SET fullname=?,gender=?,date_of_birth=?,address=?,tel=?,dep_id=?,username=?,password=?,user_type=? WHERE id=?";
             $query = $dbcon->prepare($sql);
             $query->execute(array($fullname,$gender,$date_of_birth,$address,$tel,$dep_id,$username,$password_hash,$user_type,$id));
             if($query){
@@ -93,7 +99,7 @@
         require "config.php";
         include_once("controller/app_module.php");
         $emp_id = input_data($emp_id);
-        $sql = "UPDATE users SET status = 0 WHERE id=$emp_id";
+        $sql = "UPDATE tb_user SET user_status = 0 WHERE user_id=$emp_id";
         $query = $dbcon->exec($sql);
         if($query){
             ?>
@@ -118,7 +124,7 @@
         return $query;
         // }else{
         //     $sql = "SELECT u.id,fullname,gender,date_of_birth,TRUNCATE((DATEDIFF(NOW(),date_of_birth)/365),0) 'age',tel,address,u.dep_id,d.dep_name,username,user_type
-        //     FROM users u INNER JOIN departments d ON u.dep_id = d.dep_id WHERE u.status = 1 AND u.dep_id = ? ORDER BY u.dep_id";
+        //     FROM tb_user u INNER JOIN departments d ON u.dep_id = d.dep_id WHERE u.status = 1 AND u.dep_id = ? ORDER BY u.dep_id";
         //     $query = $dbcon->prepare($sql);
         //     $query->execute(array($dep_id));
         //     return $query;
@@ -146,7 +152,7 @@
     }
     function get_emp_by_id($emp_id){
         require "config.php";
-        $sql = "SELECT * FROM users WHERE id = ?";
+        $sql = "SELECT * FROM tb_user WHERE user_id = ?";
         $query = $dbcon->prepare($sql);
         $query->execute(array($emp_id));
         return $query;
